requirements.txt
numpy
faiss-cpu
scipy
PyMuPDF
pyyaml
openai